refactor(condo): DOMA-9422 folder for common import utils
