+++ apps/condo/domains/common/utils/import/index.js
@@ -1,0 +1,5 @@
+const importDateUtils = require('@condo/domains/common/utils/import/importDate.utils')
+
+module.exports = {
+    ...importDateUtils,
+}
