+++ apps/condo/domains/meter/utils/taskSchema/AbstractMetersImporter.js
@@ -5,7 +5,7 @@ const XLSX = require('xlsx')
 const { getLogger } = require('@open-condo/keystone/logging')
 const { i18n } = require('@open-condo/locales/loader')
 
-const { clearDateStr, isDateStrValid, tryToISO } = require('@condo/domains/common/utils/importDate.utils')
+const { clearDateStr, isDateStrValid, tryToISO } = require('@condo/domains/common/utils/import/importDate.utils')
 const { IMPORT_CONDO_METER_READING_SOURCE_ID } = require('@condo/domains/meter/constants/constants')
 const { DATE_FIELD_PATHS } = require('@condo/domains/meter/constants/registerMetersReadingsService')
